requirements.txt
pygame>=2.5.0
PyOpenGL>=3.1.6
PyOpenGL_accelerate>=3.1.6
numpy>=1.21.0
pillow>=9.0.0